# swagger_server/models/create_data_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.create_data_request_vehicle import CreateDataRequestVehicle  # noqa: F401,E501
from swagger_server import util


class CreateDataRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, vehicle: CreateDataRequestVehicle=None, date_time: str=None):  # noqa: E501
        """CreateDataRequest - a model defined in Swagger

        :param vehicle: The vehicle of this CreateDataRequest.  # noqa: E501
        :type vehicle: CreateDataRequestVehicle
        :param date_time: The date_time of this CreateDataRequest.  # noqa: E501
        :type date_time: str
        """
        self.swagger_types = {
            'vehicle': CreateDataRequestVehicle,
            'date_time': str
        }

        self.attribute_map = {
            'vehicle': 'vehicle',
            'date_time': 'dateTime'
        }
        self._vehicle = vehicle
        self._date_time = date_time

    @classmethod
    def from_dict(cls, dikt) -> 'CreateDataRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateDataRequest of this CreateDataRequest.  # noqa: E501
        :rtype: CreateDataRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def vehicle(self) -> CreateDataRequestVehicle:
        """Gets the vehicle of this CreateDataRequest.


        :return: The vehicle of this CreateDataRequest.
        :rtype: CreateDataRequestVehicle
        """
        return self._vehicle

    @vehicle.setter
    def vehicle(self, vehicle: CreateDataRequestVehicle):
        """Sets the vehicle of this CreateDataRequest.


        :param vehicle: The vehicle of this CreateDataRequest.
        :type vehicle: CreateDataRequestVehicle
        """
        if vehicle is None:
            raise ValueError("Invalid value for `vehicle`, must not be `None`")  # noqa: E501

        self._vehicle = vehicle

    @property
    def date_time(self) -> str:
        """Gets the date_time of this CreateDataRequest.

        Date/Time  # noqa: E501

        :return: The date_time of this CreateDataRequest.
        :rtype: str
        """
        return self._date_time

    @date_time.setter
    def date_time(self, date_time: str):
        """Sets the date_time of this CreateDataRequest.

        Date/Time  # noqa: E501

        :param date_time: The date_time of this CreateDataRequest.
        :type date_time: str
        """
        if date_time is None:
            raise ValueError("Invalid value for `date_time`, must not be `None`")  # noqa: E501

        self._date_time = date_time