# swagger_server/models/create_data_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
import re  # noqa: F401,E501
from swagger_server import util


class CreateDataResponse(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, data_id: str=None):  # noqa: E501
        """CreateDataResponse - a model defined in Swagger

        :param data_id: The data_id of this CreateDataResponse.  # noqa: E501
        :type data_id: str
        """
        self.swagger_types = {
            'data_id': str
        }

        self.attribute_map = {
            'data_id': 'dataId'
        }
        self._data_id = data_id

    @classmethod
    def from_dict(cls, dikt) -> 'CreateDataResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateDataResponse of this CreateDataResponse.  # noqa: E501
        :rtype: CreateDataResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def data_id(self) -> str:
        """Gets the data_id of this CreateDataResponse.


        :return: The data_id of this CreateDataResponse.
        :rtype: str
        """
        return self._data_id

    @data_id.setter
    def data_id(self, data_id: str):
        """Sets the data_id of this CreateDataResponse.


        :param data_id: The data_id of this CreateDataResponse.
        :type data_id: str
        """
        if data_id is None:
            raise ValueError("Invalid value for `data_id`, must not be `None`")  # noqa: E501

        self._data_id = data_id