# swagger_server/models/update_data_request.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server.models.update_data_request_vehicle import UpdateDataRequestVehicle  # noqa: F401,E501
from swagger_server import util


class UpdateDataRequest(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, vehicle: UpdateDataRequestVehicle=None, date_time: str=None):  # noqa: E501
        """UpdateDataRequest - a model defined in Swagger

        :param vehicle: The vehicle of this UpdateDataRequest.  # noqa: E501
        :type vehicle: UpdateDataRequestVehicle
        :param date_time: The date_time of this UpdateDataRequest.  # noqa: E501
        :type date_time: str
        """
        self.swagger_types = {
            'vehicle': UpdateDataRequestVehicle,
            'date_time': str
        }

        self.attribute_map = {
            'vehicle': 'vehicle',
            'date_time': 'dateTime'
        }
        self._vehicle = vehicle
        self._date_time = date_time

    @classmethod
    def from_dict(cls, dikt) -> 'UpdateDataRequest':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The UpdateDataRequest of this UpdateDataRequest.  # noqa: E501
        :rtype: UpdateDataRequest
        """
        return util.deserialize_model(dikt, cls)

    @property
    def vehicle(self) -> UpdateDataRequestVehicle:
        """Gets the vehicle of this UpdateDataRequest.


        :return: The vehicle of this UpdateDataRequest.
        :rtype: UpdateDataRequestVehicle
        """
        return self._vehicle

    @vehicle.setter
    def vehicle(self, vehicle: UpdateDataRequestVehicle):
        """Sets the vehicle of this UpdateDataRequest.


        :param vehicle: The vehicle of this UpdateDataRequest.
        :type vehicle: UpdateDataRequestVehicle
        """

        self._vehicle = vehicle

    @property
    def date_time(self) -> str:
        """Gets the date_time of this UpdateDataRequest.

        Date/Time  # noqa: E501

        :return: The date_time of this UpdateDataRequest.
        :rtype: str
        """
        return self._date_time

    @date_time.setter
    def date_time(self, date_time: str):
        """Sets the date_time of this UpdateDataRequest.

        Date/Time  # noqa: E501

        :param date_time: The date_time of this UpdateDataRequest.
        :type date_time: str
        """

        self._date_time = date_time